first run emissionRate = float

+++ source/game/entities/game_particle.h
@@ -208,7 +208,7 @@ public:
 private:
 	Projectile *prevParticleSystem;
 
-	int emissionRateFade;
+	float emissionRateFade;
 	float verticalSpreadA;
 	float verticalSpreadB;
 	float horizontalSpreadA;
@@ -222,7 +222,7 @@ public:
 	virtual void initParticle(Particle *p, int particleIndex) override;
 	virtual void updateParticle(Particle *p) override;
 
-	void setEmissionRateFade(int emissionRateFade)		{this->emissionRateFade = emissionRateFade;}
+	void setEmissionRateFade(float emissionRateFade)	{this->emissionRateFade = emissionRateFade;}
 	void setVerticalSpreadA(float verticalSpreadA)		{this->verticalSpreadA = verticalSpreadA;}
 	void setVerticalSpreadB(float verticalSpreadB)		{this->verticalSpreadB = verticalSpreadB;}
 	void setHorizontalSpreadA(float horizontalSpreadA)	{this->horizontalSpreadA = horizontalSpreadA;}
